daksh_backend/requirements/local.txt
APScheduler==3.9.1
asgiref==3.5.0
aspose-cells==22.3.0
backports.zoneinfo==0.2.1
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.4
colorama==0.4.4
configobj==5.0.6
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==2.2
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-import-export==2.7.1
django-jsonfield-backport==1.0.4
django-rest-swagger==2.1.2
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.6.0
#fitz==0.0.1.dev2
fonttools==4.30.0
future==0.18.2
httplib2==0.20.4
idna==2.6
isodate==0.6.1
itypes==1.1.0
Jinja2==2.10
JPype1==1.3.0
kiwisolver==1.3.2
lxml==4.8.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.7.1
nibabel==3.2.2
nipype==1.7.0
numpy==1.22.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pathlib==1.0.1
pdfkit==1.0.0
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
prov==2.0.0
pydot==1.4.2
PyJWT==1.5.3
#PyMuPDF==1.19.6
PyMuPDF==1.16.14
#PyMuPDF==1.18.4
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.4
PyYAML==6.0
rdflib==6.1.1
requests==2.27.1
schedule==1.1.0
scipy==1.8.0
seaborn==0.11.2
simplejson==3.13.2
six==1.16.0
sns==0.1
sqlparse==0.2.4
tablib==3.2.0
traits==6.3.2
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==20.13.0
virtualenv-clone==0.5.7
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
